+++ src/browser/proxy.ts
@@ -7,9 +7,12 @@
 //         clipboard.on('success', onCopied);
 //     }
 
-//     function onCopied(e: ClipboardEvent) {
-//         e.clearSelection();
-//     }
+    // function onCopied(e: ClipboardEvent) {
+    //     let prevLabel = $(e.trigger).attr('aria-label');
+    //     $(e.trigger).attr('aria-label', 'Copied');
+    //     setTimeout(function () { $(e.trigger).attr('aria-label', prevLabel); }, 1000);
+    //     e.clearSelection();
+    // }
 
 //     $(document).ready(() => {
 //         initializeClipboard();