+++ src/browser/svgGenerator.ts
@@ -81,8 +81,8 @@
 //             return new Date(value);
 //         }
 
-//         return value;
-//     };
+    //     return value;
+    // }
 
 //     function drawGitGraph(startAt: number, logEntryHeight: number = 60.8, entries: LogEntry[]) {
 //         // Draw the graph