+++ src/extension.ts
@@ -4,11 +4,15 @@ import * as vscode from 'vscode';
 import * as fileHistory from './commands/fileHistory';
 import * as lineHistory from './commands/lineHistory';
 import * as logViewer from './logViewer/logViewer';
+import * as commitViewer from './commitViewer/main';
+import * as commitComparer from './commitCompare/main';
 
 // this method is called when your extension is activated
 // your extension is activated the very first time the command is executed
 export async function activate(context: vscode.ExtensionContext): Promise<any> {
     fileHistory.activate(context);
     lineHistory.activate(context);
+    commitViewer.activate(context, logViewer.getGitRepoPath);
     logViewer.activate(context);
+    commitComparer.activate(context, logViewer.getGitRepoPath);
 }